Copy May2018 ballot content into this repo
Due to how the May2018 ballot was constructed (with the entire site rather than just a single specification Markdown file), it is necessary to copy the built contents of that ballot here.

All future ballots will be treated like releases where the canonical content will be retrieved at build time.