Retrieve canonical release content during build
During a build of the site, the canonical release content will be retrieved and built alongside the side. This allows the canonical content to live in its location and not be copied/duplicated into this repository.

As part of this, relative references were used in links so that the site works locally the same as when deployed to production.

Instructions were added to document how releases & ballots should be handled in the future when building this site.