manually copied html from published site
The ballot 1.0 spec page was missing from Kevin's commit (because he added an instruction to ignore /1.0 folders in .gitignore and got tripped up). I downloaded the html for the 2018May\specification\1.0\index.html from https://cds-hooks.hl7.org/ballots/2018May/specification/1.0/ page, but them had to muck with the names of the dynamically named css and js  assets. I also commented out the html of a duplicative 1.0 menu in this same file. I don't understand why that was necessary -- something to do with mkdocs table of contents vs menu.